requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
python-dotenv
sqlalchemy
psycopg2-binary
xgboost
imbalanced-learn